+++ js/api.js
@@ -1,11 +1,3 @@
-const getData = (onSuccess) => {
-  fetch('https://22.javascript.pages.academy/kekstagram')
-    .then((response) => response.json())
-    .then((wizards) => {
-      onSuccess(wizards);
-    });
-};
-
 const sendData = (onSuccess, onFail, body) => {
   fetch(
     'https://22.javascript.pages.academy/kekstagram',
@@ -18,12 +10,39 @@ const sendData = (onSuccess, onFail, body) => {
       if (response.ok) {
         onSuccess();
       } else {
-        onFail('Не удалось отправить форму. Попробуйте ещё раз');
+        onFail();
       }
     })
     .catch(() => {
-      onFail('Не удалось отправить форму. Попробуйте ещё раз');
+      onFail();
     });
 };
 
-export {getData, sendData};
+export {sendData};
+
+
+
+/*
+ Если отправка данных прошла успешно, показывается соответствующее сообщение.
+ Разметку сообщения, которая находится блоке #success внутри шаблона template,
+  нужно разместить в <main>. Сообщение должно исчезать после нажатия на кнопку
+  .success__button, по нажатию на клавишу Esc и по клику на произвольную область
+  экрана за пределами блока с сообщением.
+
+
+3.5. Если при отправке данных произошла ошибка запроса, нужно показать
+ соответствующее сообщение. Разметку сообщения, которая находится блоке #error
+  внутри шаблона template, нужно разместить в <main>.
+  Сообщение должно исчезать
+  после нажатия на кнопки .error__button, по нажатию на клавишу Esc и по клику на
+  произвольную область экрана за пределами блока с сообщением.
+
+
+3.6. Нажатие на кнопку #upload-cancel приводит к закрытию формы и возвращению
+ всех данных и контрола фильтра к исходному состоянию (описано в пункте 3.3).
+  Поле загрузки фотографии, стилизованное под букву «О» в логотипе, очищается.
+
+
+
+*/
+
